requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
redis==5.0.8
httpx==0.27.2
